third_party/WebKit/LayoutTests/http/tests/devtools/profiler/heap-snapshot-summary-search-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests search in Summary view of detailed heap snapshots.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testSearch

Profiler was disabled.
